Fixes and improvements to authors’ notes commands and other minor fixes
Add \authorsnotefootnotemark to use with \authorsnotetext, so that both can
be disabled together.

In Chapter 4, move the note about the value of the Galleon into an
\authorsnotetext, so that it is always printed.

Remove \footnotemark commands for authors’ notes that are now in the new
environment at the top of the chapter.

Make chapterOpeningAuthorNote a comment environment by default, commenting
out its “real” definition.

Make some double quote marks curly in a newly re-added author’s note.

Add a missing full stop to a newly re-added author’s note.