Chapter 85: Improve hyphenation in a long compound word
The obvious solution to allowing discretionary hyphenation within an
already-hyphenated word (extdash with \-/) does not work, I am guessing
because of the customization we do of XeTeX character classes in
hp-format.tex (search for \dashClass).

So, use the hyphenat package, and its \hyp command.

Annoyingly, while this improves the spacing of the offending line, it still
leaves an over-large space.

See issue #176.